# apps/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  notification_channel.CreateUpdateDto:
    properties:
      active:
        type: boolean
      config:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
  notification_channel.Model:
    properties:
      active:
        type: boolean
      config:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  notification_channel.PartialUpdateDto:
    properties:
      active:
        type: boolean
      config:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
...
  utils.APIError:
    properties:
      message:
        type: string
      status:
        type: string
    required:
    - message
    - status
    type: object
  utils.ApiResponse-any:
    properties:
      data: {}
      message:
        type: string
    required:
    - data
    - message
    type: object
...
  utils.ApiResponse-array_notification_channel_Model:
    properties:
      data:
        items:
          $ref: '#/definitions/notification_channel.Model'
        type: array
      message:
        type: string
    required:
    - data
    - message
    type: object
...
  utils.ApiResponse-notification_channel_Model:
    properties:
      data:
        $ref: '#/definitions/notification_channel.Model'
      message:
        type: string
    required:
    - data
    - message
    type: object
...
info:
  contact: {}
  title: Pulsewatch API
paths:
...
  /notification-channels:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-array_notification_channel_Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: Get notification channels
      tags:
      - Notification channels
    post:
      consumes:
      - application/json
      parameters:
      - description: Notification object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/notification_channel.CreateUpdateDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse-notification_channel_Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: Create notification channel
      tags:
      - Notification channels
  /notification-channels/{id}:
    delete:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - BearerAuth: []
      summary: Delete notification channel
      tags:
      - Notification channels
    get:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-notification_channel_Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - BearerAuth: []
      summary: Get notification channel by ID
      tags:
      - Notification channels
    patch:
      consumes:
      - application/json
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      - description: Notification object
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/notification_channel.PartialUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-notification_channel_Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - BearerAuth: []
      summary: Update notification channel
      tags:
      - Notification channels
    put:
      consumes:
      - application/json
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      - description: Notification object
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/notification_channel.CreateUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-notification_channel_Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - BearerAuth: []
      summary: Update notification channel
      tags:
      - Notification channels
  /notification-channels/test:
    post:
      consumes:
      - application/json
      parameters:
      - description: Notification object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/notification_channel.CreateUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: Test notification channel
      tags:
      - Notification channels
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication (pk_ prefix format)
    in: header
    name: X-API-Key
    type: apiKey
  JwtAuth:
    description: JWT token authentication (Bearer token format)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"